Ele Boss Fixes, Shifter Fixes, Lich and Lycan updates
So I did adjustments to the Elemental Boss (Added 5/- Divine Resist, Darkvision, Darkness immunity to the boss, I adjusted one of the area spawners that was causing units to spawn into cliffs, and I changed the exit portal color to stand out more)

I boosted the Lich drop rate to 10% Divine, 20% unique, and 3 Epic drops

Fixed the Lich helmet bug (It was causing it to summon a stream of zombies once used, this has to do with the fact there is an execute script tied to items in the equ and unequ mod files. I just added an exception to the code for the helmet)

Shifter crashing should be fixed - We need to test ( Forcing players to polymorph via the radial menu and adding in a 1 round delay. So once you polymorph you have a 1 round delay in repolymophing. This should add a bit of a stagger if someone tries to spam somehow.)

Lycan should be fixed? Maybe. I moved the functions into the script itself so we just gotta see if it works on the test, push it live and see.

I also made Frostspear's treasure blue to match the rest of the loot